add readme and refactor utils in operator dir
