+++ tests/unit/operator/test_operator_utils.py
@@ -4,12 +4,14 @@ from types import SimpleNamespace
 
 import pytest
 
+import bentoctl.operator.utils
+import bentoctl.operator.utils.git
 from bentoctl.operator import utils as operator_utils
 
 
 def test_get_bentoctl_home(tmp_path):
     os.environ["BENTOCTL_HOME"] = str(tmp_path)
-    bentoctl_home = operator_utils._get_bentoctl_home()
+    bentoctl_home = bentoctl.operator.utils._get_bentoctl_home()
     assert bentoctl_home == tmp_path
     assert (tmp_path / "operators").exists()
 
@@ -24,7 +26,7 @@ def test_get_bentoctl_home(tmp_path):
     ],
 )
 def test_is_git_link(git_link, truth):
-    assert operator_utils._is_git_link(git_link) is truth
+    assert bentoctl.operator.utils.git._is_git_link(git_link) is truth
 
 
 @pytest.mark.parametrize(
@@ -36,19 +38,19 @@ def test_is_git_link(git_link, truth):
     ],
 )
 def test_is_github_repo(github_repo, truth):
-    assert operator_utils._is_github_repo(github_repo) is truth
+    assert bentoctl.operator.utils.git._is_github_repo(github_repo) is truth
 
 
 @pytest.mark.parametrize(
     "official_op, truth", [("aws-lambda", True), ("testop", False)]
 )
 def test_is_official_operator(official_op, truth):
-    assert operator_utils._is_official_operator(official_op) is truth
+    assert bentoctl.operator.utils._is_official_operator(official_op) is truth
 
 
 def test_get_operator_dir_path(tmp_path):
     os.environ["BENTOCTL_HOME"] = str(tmp_path)
-    op_dir = operator_utils._get_operator_dir_path(operator_name="test_operator")
+    op_dir = bentoctl.operator.utils._get_operator_dir_path(operator_name="test_operator")
     assert op_dir == str(tmp_path / "operators" / "test_operator")
 
 
@@ -68,11 +70,11 @@ class PatchedRepo:
 
 def test_clone_git_repo(monkeypatch):
     monkeypatch.setattr(operator_utils, "Repo", PatchedRepo)
-    repo_path = operator_utils._clone_git_repo("git_url")
+    repo_path = bentoctl.operator.utils.git._clone_git_repo("git_url")
     assert os.path.exists(repo_path)
 
     # checkout with branch
-    repo_path = operator_utils._clone_git_repo("git_url", branch="test")
+    repo_path = bentoctl.operator.utils.git._clone_git_repo("git_url", branch="test")
     assert os.path.exists(repo_path)
     assert os.path.exists(os.path.join(repo_path, "test"))
 
@@ -88,7 +90,7 @@ def test_clone_git_repo(monkeypatch):
 def test_fetch_github_info(github_link, info, raise_error):
     if raise_error:
         with pytest.raises(ValueError):
-            operator_utils._fetch_github_info(github_link)
+            bentoctl.operator.utils.git._fetch_github_info(github_link)
     else:
-        returned_info = operator_utils._fetch_github_info(github_link)
+        returned_info = bentoctl.operator.utils.git._fetch_github_info(github_link)
         assert returned_info == info
